# Conflicts:
#	src/main/kotlin/service/user/UserService.kt

+++ src/main/kotlin/service/store/data/Model.kt
@@ -1,8 +1,10 @@
 package service.store.data
 
+import java.math.BigDecimal
+
 data class ThinPokemon(
     val id: Int,
     val name: String,
-    var xp: Long,
+    var xp: BigDecimal,
     val sprite: String?
 )
